Add ability to reset request headers on client attribute
Because of the way SendGridAPIClient and python_http_client are implemented, request headers are always cached. This becomes a problem when executing a sequence of API requests that require varying request header values using the same SendGridAPIClient instance.
This method allows the caller to reset the request headers on the client instance to their default.