Remove signs of Python 3.2 and Python 3.3
This removes Python 3.2 and Python 3.3 from the Trove classifiers for the project, so they no longer show up on PyPI as supported versions. This matches the versions tested in Travis CI as well as the versions stated in the README.

This removes the requirement handling for Python 3.0 and Python 3.1, which haven't been supported for quite some time.

This also blocks installation of the package of the older versions of Python that are no longer supported, so Pip won't even attempt to install it.