Update webhook auth docstring
The verify_signature method returns False if the event payload is decoded
using utf-8 if special characters exist. The payload must be decoded
using latin-1 to have it not fail on what should be successful
authorizations. It is not clear from the current docstring what encoding
to use when decoding the bytes payload