Fix #988: Remove async example
The async example never actually worked in an async way. It tricked
users into thinking they were using async code but in reality the
sendgrid_client.send call (which is what was waiting on I/O) was still
blocking.

In addition, asyncio.async (used in the example) was deprecated in
Python 3.7 and removed in Python 3.10.